+++ crates/node/tests/globals_test.rs
@@ -1,3 +1,0 @@
-use proto_pdk_test_utils::*;
-
-generate_globals_test!("node-test", "prettier");