+++ crates/node/tests/versions_test.rs
@@ -6,8 +6,8 @@ generate_resolve_versions_tests!("node-test", {
     "10.1" => "10.1.0",
     "lts-gallium" => "16.20.2",
     "lts/fermium" => "14.21.3",
-    "stable" => "20.11.0",
-    "node" => "21.6.1",
+    "stable" => "20.11.1",
+    "node" => "21.6.2",
 });
 
 #[test]